+++ packages/CVMTests/src/package-test.ts
@@ -4,6 +4,7 @@ import * as DG from 'datagrok-api/dg';
 import './shell/ml';
 import './scripts/scripts_tests';
 import './docker/docker';
+import './files/files'
 // import './gui/dialogs'; To fix!
 
 import {runTests, tests, TestContext} from '@datagrok-libraries/utils/src/test';