+++ packages/Charts/src/viewers/radar/radar-viewer.ts
@@ -3,12 +3,16 @@ import * as ui from 'datagrok-api/ui';
 import * as grok from 'datagrok-api/grok';
 
 import * as echarts from 'echarts';
-import {option} from './constants';
+import {MAXIMUM_SERIES_NUMBER, option} from './constants';
 import {StringUtils} from '@datagrok-libraries/utils/src/string-utils';
+import { EChartViewer } from '../echart/echart-viewer';
+
+import '../../../css/radar-viewer.css';
 
 type MinimalIndicator = '1' | '5' | '10' | '25';
 type MaximumIndicator = '75' | '90' | '95' | '99';
 const ERROR_CLASS = 'd4-viewer-error';
+const WARNING_CLASS = 'radar-warning';
 
 // Based on this example: https://echarts.apache.org/examples/en/editor.html?c=radar
 @grok.decorators.viewer({
@@ -16,20 +20,20 @@ const ERROR_CLASS = 'd4-viewer-error';
   description: 'Creates a radar viewer',
   icon: 'icons/radar-viewer.svg',
 })
-export class RadarViewer extends DG.JsViewer {
+export class RadarViewer extends EChartViewer {
   get type(): string {return 'RadarViewer';}
 
   chart: echarts.ECharts;
   min: MinimalIndicator;
   max: MaximumIndicator;
   showOnlyCurrentRow: boolean;
-  showAllRows: boolean;
   showTooltip: boolean;
   showMin: boolean;
   showMax: boolean;
   showValues: boolean;
   backgroundMinColor: number;
   backgroundMaxColor: number;
+  currentRowColor: number;
   valuesColumnNames: string[];
   columns: DG.Column[] = [];
   title: string;
@@ -37,22 +41,22 @@ export class RadarViewer extends DG.JsViewer {
 
   constructor() {
     super();
+    this.rowSource = 'CurrentRow';
     this.title = this.string('title', 'Radar');
     this.min = <MinimalIndicator> this.string('min', '5', { choices: ['1', '5', '10', '25'],
       description: 'Minimum percentile value (indicated as dark blue area)' });
     this.max = <MaximumIndicator> this.string('max', '95', { choices: ['75', '90', '95', '99'],
       description: 'Maximum percentile value (indicated as light blue area)' });
     this.showOnlyCurrentRow = this.bool('showOnlyCurrentRow', true, {description: 'Hides max and min values'});
-    this.showAllRows = this.bool('showAllRows', false);
     this.showTooltip = this.bool('showTooltip', true);
     this.backgroundMinColor = this.int('backgroundMinColor', 0xFFB0D7FF);
     this.backgroundMaxColor = this.int('backgroundMaxColor', 0xFFBCE2F5);
-    this.showMin = this.bool('showMin', true);
-    this.showMax = this.bool('showMax', true);
+    this.currentRowColor = this.int('currentRowColor', 0xFFBAFFBA);
+    this.showMin = this.bool('showMin', false);
+    this.showMax = this.bool('showMax', false);
     this.showValues = this.bool('showValues', true);
     this.valuesColumnNames = this.addProperty('valuesColumnNames', DG.TYPE.COLUMN_LIST, null,
       {columnTypeFilter: DG.TYPE.NUMERICAL});
-    this.addRowSourceAndFormula();
     
     const chartDiv = ui.div([], { style: { position: 'absolute', left: '0', right: '0', top: '0', bottom: '0'}} );
     this.root.appendChild(chartDiv);
@@ -101,14 +105,6 @@ export class RadarViewer extends DG.JsViewer {
     this.helpUrl = 'https://datagrok.ai/help/visualize/viewers/radar';
   }
 
-  initChartEventListeners() {
-    this.dataFrame.onRowsFiltered.subscribe((_) => {
-      if (this.dataFrame) {
-        this.render();
-      }
-    });
-  }
-
   getTooltip(params: any): string | null {
     if (params.componentType === 'series') {
       if (params.seriesIndex === 2) {
@@ -123,19 +119,9 @@ export class RadarViewer extends DG.JsViewer {
 
   onTableAttached() {
     this.init();
-    this.initChartEventListeners();
     this.valuesColumnNames = Array.from(this.dataFrame.columns.numerical)
       .map((c: DG.Column) => c.name);
-    this.subs.push(this.dataFrame.selection.onChanged.subscribe((_) => this.render()));
-    this.subs.push(this.dataFrame.filter.onChanged.subscribe((_) => this.render()));
-    this.subs.push(this.dataFrame.onCurrentRowChanged.subscribe((_) => {
-      this.updateRow();
-      this.chart.setOption(option);
-    }));
-    this.subs.push(this.dataFrame.onColumnsRemoved.subscribe((_) => {
-      this.init();
-      this.chart.setOption(option);
-    }));
+    this.subs.push(this.dataFrame.onCurrentRowChanged.subscribe((_) => this.render()));
     this.render();
   }
 
@@ -160,30 +146,58 @@ export class RadarViewer extends DG.JsViewer {
   
     if (this.showMax)
       this.updateMax();
-  
+
     if (this.showOnlyCurrentRow)
       this.updateRow();
-  
-    if (this.showAllRows) {
-      option.series[2].data = [];
-      for (let i = 0; i < this.dataFrame.rowCount; i++) {
-        option.series[2].data.push({
-          value: this.columns.map((c) => {
-            const value = Number(c.get(i));
-            return value !== -2147483648 ? value : 0;
-          }),
+
+    const currentRowIndex = this.dataFrame.currentRowIdx ?? 0;
+    const currentIn = this.filter.get(currentRowIndex);
+    
+    if (!currentIn || !this.showOnlyCurrentRow) {
+      const filter = this.formulaFilter ? this.filter.clone().invert() : this.filter;
+      const seriesData = [];
+      let colorIndex = 0;
+
+      for (let i = 0; i < filter.length && seriesData.length < MAXIMUM_SERIES_NUMBER; i++) {
+        if (!filter.get(i)) continue;
+        if (i === currentRowIndex && currentIn) continue;
+        
+        const value = this.columns.map((c) => {
+          if (c.type === 'datetime')
+            return this.getDate(c, c.getRawData()[i]);
+          const value = Number(c.get(i));
+          return value !== -2147483648 ? value : 0;
+        });
+
+        const colors = DG.Color.categoricalPalette;
+        const color = colors[colorIndex % colors.length];
+        ++colorIndex;
+
+        seriesData.push({
+          value: value,
           name: `row ${i + 1}`,
+          itemStyle: { color: DG.Color.toHtml(color) },
           label: { show: this.showValues, formatter: (params: any) => StringUtils.formatNumber(params.value) as string },
         });
       }
+      
+      option.series[2].data = seriesData;
+      if (currentIn && !this.showOnlyCurrentRow)
+        this.updateRow();
     }
-    
-    option.legend.show = !this.showAllRows;
+
+    option.legend.show = !(this.filter.length > 1);
     option.silent = !this.showTooltip;
-  }  
+  }
+
+  getDate(c: DG.Column, value: number) {
+    const date = this.getYearFromDate(value);
+    const isRight = date >= this.getYearFromDate(c.min) && date <= this.getYearFromDate(c.max);
+    return isRight ? date : this.getYearFromDate(c.min);
+  }
 
   getYearFromDate(value: number) {
-    return new Date(Math.floor(value) / 1000).getFullYear()
+    return new Date(Math.floor(value) / 1000).getFullYear();
   }
 
   updateMin() {
@@ -220,10 +234,10 @@ export class RadarViewer extends DG.JsViewer {
 
   updateRow() {
     const currentRow = Math.max(this.dataFrame.currentRowIdx, 0);
-    option.series[2].data[0] = {
+    option.series[2].data.unshift({
       value: this.columns.map((c) => {
         if (c.type === 'datetime')
-          return this.getYearFromDate(c.getRawData()[currentRow]);
+          return this.getDate(c, c.getRawData()[currentRow]);
         const value = Number(c.get(currentRow));
         return value != -2147483648 ? value : 0;
       }),
@@ -231,11 +245,11 @@ export class RadarViewer extends DG.JsViewer {
       lineStyle: {
         width: 2,
         type: 'dashed',
-        color: 'rgba(66, 135, 204, 0.8)',
+        color: DG.Color.toHtml(this.currentRowColor),
       },
       symbolSize: 6,
       itemStyle: {
-        color: 'rgba(66, 135, 204, 0.8)',
+        color: DG.Color.toHtml(this.currentRowColor),
       },
       label: {
         show: this.showValues,
@@ -243,7 +257,7 @@ export class RadarViewer extends DG.JsViewer {
           return StringUtils.formatNumber(params.value) as string;
         },
       },
-    };
+    });
   }
 
   clearData(indexes: number[]) {
@@ -266,29 +280,35 @@ export class RadarViewer extends DG.JsViewer {
     return columns;
   }
 
-  _showErrorMessage(msg: string) {this.root.appendChild(ui.divText(msg, ERROR_CLASS));}
+  _showMessage(msg: string, className: string) {this.root.appendChild(ui.divText(msg, className));}
 
-  _removeErrorMessage() {
-    const divTextElement = this.root.getElementsByClassName(ERROR_CLASS)[0];
+  _removeMessage(className: string) {
+    const divTextElement = this.root.getElementsByClassName(className)[0];
     if (divTextElement)
       this.root.removeChild(divTextElement);
   }
 
   render() {
     if (this.valuesColumnNames.length < 1) {
-      this._showErrorMessage('The Radar viewer requires a minimum of 1 numerical column.');
+      this._showMessage('The Radar viewer requires a minimum of 1 numerical column.', ERROR_CLASS);
       return;
     }
-    this._removeErrorMessage();
+
+    this._removeMessage(WARNING_CLASS);
+    this._removeMessage(ERROR_CLASS);
     this.getSeriesData();
     this.chart.setOption(option);
+
+    if ((!this.showOnlyCurrentRow && this.filter.trueCount > MAXIMUM_SERIES_NUMBER) ||
+      (this.showOnlyCurrentRow && option.series[2].data.length > 1))
+      this._showMessage('Only first 100 rows are shown', WARNING_CLASS);
   }
 
   /* Going to be replaced with perc in stats */
   getQuantile(columns: DG.Column<any>[], percent: number) {
     const result = [];
     for (const c of columns) {
-      const datetime = c.getRawData().map((value: number) => this.getYearFromDate(value));
+      const datetime = c.getRawData().map((value: number) => this.getDate(c, value));
       const values = c.type === 'datetime' ? datetime : Array.from(c.values());
       const isValidValue = (value: number) => typeof value === 'bigint' 
         ? value !== BigInt('-2147483648')