+++ packages/ODEs/src/package.ts
@@ -58,7 +58,8 @@ export async function ivpFileHandler(content: string) {
 }
 
 //name: previewIvp
-//tags: fileViewer, fileViewer-ivp
+//tags: fileViewer
+//meta.fileViewer: ivp
 //input: file file
 //output: view preview
 export async function previewIvp(file: DG.FileInfo): Promise<DG.View> {