+++ packages/UsageAnalysis/src/tabs/reports.ts
@@ -13,11 +13,17 @@ const filtersStyle = {
 
 export class ReportsView extends UaView {
   currentFilterGroup: DG.FilterGroup | null;
+  private filters: HTMLDivElement = ui.box();
 
   constructor(uaToolbox: UaToolbox) {
     super(uaToolbox);
     this.name = 'Reports';
     this.currentFilterGroup = null;
+    this.filters.style.maxWidth = '250px';
+    this.ribbonMenu = DG.Menu.create();
+    this.ribbonMenu.item('Text', () => {
+      console.log('From ribbon');
+    });
   }
 
   async initViewers(): Promise<void> {
@@ -30,49 +36,23 @@ export class ReportsView extends UaView {
       };
     });
 
-    const filters = ui.box();
-    filters.style.maxWidth = '250px';
-
     const reportsViewer = new UaFilterableQueryViewer({
       filterSubscription: this.uaToolbox.filterStream,
       name: 'User reports',
       queryName: 'UserReports',
       createViewer: (t: DG.DataFrame) => {
         const viewer = DG.Viewer.grid(t, {
-          'showColumnLabels': false,
-          'showRowHeader': false,
-          'showColumnGridlines': false,
-          'allowRowSelection': false,
-          'allowColumnSelection': false,
-          'allowBlockSelection': false,
-          'showCurrentCellOutline': false,
-          'defaultCellFont': '13px monospace',
+          'defaultCellFont': '13px monospace'
         });
-
-        const filters_ = DG.Viewer.filters(t, filtersStyle);
-        this.currentFilterGroup = new DG.FilterGroup(filters_.dart);
-        if (ViewHandler.getCurrentView().name === 'Reports' && ViewHandler.getInstance().getSearchParameters().has('report-number')) {
-          let reportNumber = ViewHandler.getInstance().getSearchParameters().get('report-number');
-          if (reportNumber)
-            this.updateFilter(reportNumber);
-        }
-        filters.append(filters_.root);
+        this.reloadFilter(t);
         viewer.onBeforeDrawContent.subscribe(() => {
-          viewer.columns.setOrder(['report_id', 'reporter', 'report_time', 'description', 'same_errors_count', 'error', 'error_stack_trace']);
+          viewer.columns.setOrder(['report_number', 'reporter', 'description', 'same_errors_count', 'error', 'error_stack_trace', 'report_time', 'report_id']);
           viewer.col('reporter')!.cellType = 'html';
-          viewer.col('reporter')!.width = 30;
-          viewer.col('report_id')!.cellType = 'html';
-          viewer.col('report_id')!.width = 20;
-          viewer.col('report_number')!.visible = false;
+          viewer.col('error_stack_trace_hash')!.visible = false;
         });
 
         viewer.onCellPrepare(async function(gc) {
-          if (gc.gridColumn.name === 'report_time') {
-            gc.style.textColor = 0xFFB8BAC0;
-            gc.style.font = '13px Roboto';
-          }
-
-          if (gc.gridColumn.name === 'reporter') {
+          if (gc.gridColumn.name === 'reporter' && gc.cell.value) {
             const user = users[gc.cell.value];
             const img = ui.div();
             img.style.width = '20px';
@@ -84,22 +64,13 @@ export class ReportsView extends UaView {
               img.style.backgroundImage = user.avatar.style.backgroundImage;
             else
               img.style.backgroundImage = 'url(/images/entities/grok.png);';
-            img.addEventListener('click', () => {
+            const span = ui.span([ui.span([img, ui.label(user.name)], 'grok-markup-user')], 'd4-link-label');
+            span.addEventListener('click', () => {
               grok.shell.o = user.data;
             });
-            gc.style.element = ui.tooltip.bind(img, user.name);
-          }
-
-          if (gc.gridColumn.name === 'report_id') {
-            const icon = ui.iconFA('arrow-to-bottom', () => {
-              const indicator = DG.TaskBarProgressIndicator.create('Receiving user report...');
-              //@ts-ignore
-              grok.dapi.admin.getUserReport(gc.cell.value)
-                .then((bytes: any) => DG.Utils.download(`report_${gc.cell.value}.zip`, bytes))
-                .finally(() => indicator.close());
+            gc.style.element = ui.tooltip.bind(span, () => {
+              return DG.ObjectHandler.forEntity(user.data)?.renderTooltip(user.data.dart)!;
             });
-            icon.style.marginTop = '7px';
-            gc.style.element = ui.tooltip.bind(icon, 'Download report');
           }
         });
 
@@ -120,17 +91,44 @@ export class ReportsView extends UaView {
     reportsViewer.root.classList.add('ui-panel');
     this.viewers.push(reportsViewer);
     this.root.append(ui.splitH([
-      filters,
+      this.filters,
       reportsViewer.root,
     ]));
   }
 
-  updateFilter(reportNumber: string): void {
-    this.currentFilterGroup?.updateOrAdd({
-      type: DG.FILTER_TYPE.MULTI_VALUE,
-      column: 'report_number',
-      selected: [reportNumber]}
-    );
+  async reloadViewers(): Promise<void> {
+    this.viewers = [];
+    while (this.root.hasChildNodes())
+      this.root.removeChild(this.root.lastChild!);
+    await this.initViewers();
+    for (const v of this.viewers)
+      await v.reloadViewer();
+  }
+
+  async reloadFilter(table?: DG.DataFrame) {
+    this.currentFilterGroup?.detach();
+    while (this.filters.hasChildNodes())
+      this.filters.removeChild(this.filters.lastChild!);
+    table = table ?? await this.viewers[0].dataFrame;
+    if (table) {
+      const filters_ = DG.Viewer.filters(table, filtersStyle);
+      this.currentFilterGroup = new DG.FilterGroup(filters_.dart);
+      this.filters.append(filters_.root);
+      this.updateFilter();
+    }
+  }
+
+  updateFilter(): void {
+    if (ViewHandler.getCurrentView().name === 'Reports' && ViewHandler.getInstance().getSearchParameters().has('report-number')) {
+      let reportNumber = ViewHandler.getInstance().getSearchParameters().get('report-number');
+      if (reportNumber) {
+        this.currentFilterGroup?.updateOrAdd({
+          type: DG.FILTER_TYPE.MULTI_VALUE,
+          column: 'report_number',
+          selected: [reportNumber]}
+        );
+      }
+    }
   }
 
   async showReportContextPanel(table: DG.DataFrame): Promise<void> {