+++ packages/Tutorials/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Tutorials changelog
 
+## 1.3.23 (2024-04-11)
+
+Fixed a couple of bugs in Cheminformatics tutorials
+
 ## 1.3.22 (2023-12-21)
 
 ### Features